# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: LMS Academy API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (admin/owner only)
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: search
          in: query
          schema:
            type: string
        - name: role
          in: query
          schema:
            $ref: "#/components/schemas/UserRole"
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"
  /users/{userId}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User details with enrollments and orders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDetail"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user profile
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"

    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete user (owner only)
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /users/{userId}/password:
    patch:
      operationId: changeUserPassword
      tags: [users]
      summary: Change user password (self or admin)
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # COURSES
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]

    # ENUMS
    UserRole:
      type: string
      enum: [owner, instructor, student]
...
    EnrollmentStatus:
      type: string
      enum: [active, revoked, expired]
    OrderStatus:
      type: string
      enum: [pending, paid, failed, refunded]
...
    ChangePasswordRequest:
      type: object
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
          minLength: 8
      required: [newPassword]

    # USER SCHEMAS
    UserProfile:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          $ref: "#/components/schemas/UserRole"
        avatarUrl:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, name, email, role, createdAt]
    UserDetail:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          $ref: "#/components/schemas/UserRole"
        avatarUrl:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        enrollments:
          type: array
          items:
            $ref: "#/components/schemas/Enrollment"
        orders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
      required: [id, name, email, role, createdAt, enrollments, orders]
    UpdateUserRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        role:
          $ref: "#/components/schemas/UserRole"
        avatarUrl:
          type: string
    UserListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/UserProfile"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [data, total, page, limit]

    # COURSE SCHEMAS
...
    Enrollment:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        courseId:
          type: string
        courseName:
          type: string
          nullable: true
        userName:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/EnrollmentStatus"
        enrolledAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
          nullable: true
      required: [id, userId, courseId, status, enrolledAt]
...
    Order:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        courseId:
          type: string
        courseName:
          type: string
          nullable: true
        userName:
          type: string
          nullable: true
        amount:
          type: number
        currency:
          type: string
        status:
          $ref: "#/components/schemas/OrderStatus"
        paymentGateway:
          type: string
          nullable: true
        gatewayOrderId:
          type: string
          nullable: true
        affiliateCode:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        paidAt:
          type: string
          format: date-time
          nullable: true
      required: [id, userId, courseId, amount, currency, status, createdAt]
...